Feat: Refactor component imports and update DAY constants for better structure
