--- src/components/DataTable.tsx
@@ -1,167 +1,0 @@
-"use client";
-import * as React from "react";
-
-import {
-	ColumnDef,
-	ColumnFiltersState,
-	SortingState,
-	flexRender,
-	getCoreRowModel,
-	getFilteredRowModel,
-	getPaginationRowModel,
-	getSortedRowModel,
-	useReactTable,
-} from "@tanstack/react-table";
-
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import {
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
-} from "@/components/ui/table";
-
-interface DataTableProps<TData, TValue> {
-	columns: ColumnDef<TData, TValue>[];
-	columnsRech?: string;
-	placeholderInput?: string;
-	columnsRech2?: string;
-	placeholderInput2?: string;
-	data: TData[];
-	children?: React.ReactNode;
-}
-
-export function DataTable<TData, TValue>({
-	columns,
-	columnsRech,
-	placeholderInput,
-	columnsRech2,
-	placeholderInput2,
-	data,
-	children,
-}: DataTableProps<TData, TValue>) {
-	const [sorting, setSorting] = React.useState<SortingState>([]);
-	const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
-		[]
-	);
-	const table = useReactTable({
-		data,
-		columns,
-		getCoreRowModel: getCoreRowModel(),
-		getPaginationRowModel: getPaginationRowModel(),
-		onSortingChange: setSorting,
-		getSortedRowModel: getSortedRowModel(),
-		onColumnFiltersChange: setColumnFilters,
-		getFilteredRowModel: getFilteredRowModel(),
-		state: {
-			sorting,
-			columnFilters,
-		},
-	});
-
-	return (
-		<div>
-			<div className="flex flex-row space-x-4">
-				<div className="flex items-center py-4">{children}</div>
-				<div className="flex items-center py-4">
-					{columnsRech && (
-						<Input
-							placeholder={placeholderInput}
-							value={
-								(table.getColumn(columnsRech)?.getFilterValue() as string) ?? ""
-							}
-							onChange={(event) =>
-								table.getColumn(columnsRech)?.setFilterValue(event.target.value)
-							}
-							className="max-w-sm"
-						/>
-					)}
-				</div>
-				<div className="flex items-center py-4">
-					{columnsRech2 && (
-						<Input
-							placeholder={placeholderInput2}
-							value={
-								(table.getColumn(columnsRech2)?.getFilterValue() as string) ??
-								""
-							}
-							onChange={(event) =>
-								table
-									.getColumn(columnsRech2)
-									?.setFilterValue(event.target.value)
-							}
-							className="max-w-sm"
-						/>
-					)}
-				</div>
-			</div>
-			<div className="rounded-lg border">
-				<Table>
-					<TableHeader className="bg-slate-50">
-						{table.getHeaderGroups().map((headerGroup) => (
-							<TableRow key={headerGroup.id}>
-								{headerGroup.headers.map((header) => {
-									return (
-										<TableHead key={header.id}>
-											{header.isPlaceholder
-												? null
-												: flexRender(
-														header.column.columnDef.header,
-														header.getContext()
-												  )}
-										</TableHead>
-									);
-								})}
-							</TableRow>
-						))}
-					</TableHeader>
-					<TableBody>
-						{table.getRowModel().rows?.length ? (
-							table.getRowModel().rows.map((row) => (
-								<TableRow
-									key={row.id}
-									data-state={row.getIsSelected() && "selected"}>
-									{row.getVisibleCells().map((cell) => (
-										<TableCell key={cell.id}>
-											{flexRender(
-												cell.column.columnDef.cell,
-												cell.getContext()
-											)}
-										</TableCell>
-									))}
-								</TableRow>
-							))
-						) : (
-							<TableRow>
-								<TableCell
-									colSpan={columns.length}
-									className="h-24 text-center">
-									Aucun Resultat
-								</TableCell>
-							</TableRow>
-						)}
-					</TableBody>
-				</Table>
-			</div>
-			<div className="flex items-center justify-end space-x-2 py-4">
-				<Button
-					variant="outline"
-					size="sm"
-					onClick={() => table.previousPage()}
-					disabled={!table.getCanPreviousPage()}>
-					Précédent
-				</Button>
-				<Button
-					variant="outline"
-					size="sm"
-					onClick={() => table.nextPage()}
-					disabled={!table.getCanNextPage()}>
-					Suivant
-				</Button>
-			</div>
-		</div>
-	);
-}
