--- src/lib/Matter.ts
@@ -1,9 +1,0 @@
-interface Matter {
-	name: string;
-	Abr_Matter: string;
-}
-
-interface List_Matter {
-	Level: string;
-	Matter: Matter[];
-}
