--- src/lib/Teacher.ts
@@ -1,4 +1,0 @@
-interface Teacher {
-	name: string;
-	Abr_Teacher: string;
-}
